requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
>=0.3.14
python-telegram-bot>=20.7
PyYAML>=6.0
APScheduler>=3.10.4
python-dateutil>=2.8.2
